ciri: remove conf_provider
